add checkbox to risk view rename configure file
