+++ packages/frontend/src/pages/scaling-risk/view/ScalingRiskView.stories.tsx
@@ -4,7 +4,7 @@ import React, { useEffect } from 'react'
 
 import { PageContent } from '../../../components/PageContent'
 import { Tooltip } from '../../../components/Tooltip'
-import { configureTabNavigations } from '../../../scripts/configureTabNavigations'
+import { configureTabNavigations } from '../../../scripts/configureTabs'
 import { configureTooltips } from '../../../scripts/configureTooltips'
 import { click } from '../../../utils/storybook/click'
 import { ScalingRiskView } from './ScalingRiskView'
