+++ packages/frontend/src/scripts/configureTabs.ts
