+++ packages/frontend/src/scripts/index.ts
@@ -7,7 +7,7 @@ import { configureDropdowns } from './configureDropdowns'
 import { configureExpandableContainer } from './configureExpandableContainer'
 import { configureHoverableDropdown } from './configureHoverableDropdown'
 import { configureRosetteOverlay } from './configureRosetteOverlay'
-import { configureTabNavigations } from './configureTabNavigations'
+import { configureTabNavigations } from './configureTabs'
 import { configureTooltips } from './configureTooltips'
 import { configureProjectNavigation } from './section-navigation'
